backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
opencv-python-headless==4.10.0.84
google-genai==1.14.0
firebase-admin==6.6.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.9
